navbar improved
