+++ src/app/(home)/layout.tsx
@@ -1,17 +1,16 @@
 import FooterOne from "@/components/layout/FooterOne";
 import HeaderOne from "@/components/layout/HeaderOne";
+import InstagramFeed from "@/components/layout/InstagramFeed";
+import React from "react";
 
-const Layout = ({ children }: { children: React.ReactNode }) => {
-    return (
-        <>
-            <HeaderOne />
-            <main className="xl:pt-[100px] pt-[72px]">
-                {children}
-            </main>
-            <FooterOne />
-        </>
-    );
-}
+const HomeParallaxLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <>
+      <HeaderOne variant="transparent" />
+      <main>{children}</main>
+      <FooterOne />
+    </>
+  );
+};
 
-
-export default Layout;
+export default HomeParallaxLayout;
