+++ src/app/(home)/page.tsx
@@ -1,44 +1,47 @@
-import { Metadata } from "next";
-import AboutOne from "@/components/HomeOne/AboutOne";
 import BlogOne from "@/components/HomeOne/BlogOne";
-import CategoryOne from "@/components/HomeOne/CategoryOne";
-import Clients from "@/components/HomeOne/Clients";
-import Faq from "@/components/HomeOne/Faq";
-import FeaturedPackages from "@/components/HomeOne/FeaturedPackages";
-import Hero from "@/components/HomeOne/Hero";
-import HeroBulletin from "@/components/HomeOne/HeroBulletin";
-import PackageOne from "@/components/HomeOne/PackageOne";
-import TestimonialOne from "@/components/HomeOne/TestimonialOne";
 import VideoBanner from "@/components/HomeOne/VideoBannerOne";
+import ClientsTwo from "@/components/HomeParallax/ClientsTwo";
+import DestinationTwo from "@/components/HomeParallax/DestinationTwo";
+import Features from "@/components/HomeParallax/Features";
+import HeroThree from "@/components/HomeParallax/HeroThree";
+import PackageThree from "@/components/HomeParallax/PackageThree";
+import TestimonialThree from "@/components/HomeParallax/TestimonialThree";
 import InstagramFeed from "@/components/layout/InstagramFeed";
-
+import { Metadata } from "next";
+import React from "react";
 
 export const metadata: Metadata = {
-    title: 'Home 01 | Arid - Travel & Tourism HTML/Tailwind CSS Template',
-    description: 'Welcome, Arid - Travel & Tourism HTML/Tailwind CSS Template',
-    keywords: ['tour', 'travel', 'booking', 'rental', 'nextJs', 'tailwind', 'trip', 'beach', 'holidy', 'cruise', 'vacation' ]
-}
+  title: "Aktivferien Nepal | Trekking & Abenteuerreisen im Himalaya",
+  description:
+    "Erleben Sie unvergessliche Trekking- und Abenteuerreisen in Nepal mit Aktivferien. Geführt von erfahrenen Schweizer Bergführern, entdecken Sie die Schönheit des Himalayas.",
+  keywords: [
+    "Nepal",
+    "Trekking",
+    "Himalaya",
+    "Aktivferien",
+    "Abenteuerreisen",
+    "Schweizer Bergführer",
+    "Mustang",
+    "Annapurna",
+    "Everest Base Camp",
+    "Kulturreisen",
+  ],
+};
 
+const HomeParallax = () => {
+  return (
+    <>
+      <HeroThree />
+      <PackageThree />
+      <DestinationTwo />
+      <Features />
+      <VideoBanner />
+      <ClientsTwo />
+      <TestimonialThree />
+      <BlogOne />
+      <InstagramFeed />
+    </>
+  );
+};
 
-const HomeOne = () => {
-    return (
-        <>
-            <Hero />
-            <HeroBulletin />
-            <CategoryOne />
-            <PackageOne />
-            <VideoBanner />
-            <div className="bg-gradient-to-t to-[#FFF1EC] from-white ">
-                <Clients />
-                <AboutOne />
-            </div>
-            <FeaturedPackages />
-            <TestimonialOne />
-            <Faq />
-            <BlogOne />
-            <InstagramFeed />
-        </>
-    );
-}
-
-export default HomeOne;
+export default HomeParallax;
